+++ src/datasources/alarms/query-type-handlers/alarm-trend/AlarmTrendQueryHandler.ts
@@ -1,14 +1,14 @@
 import { DataFrameDTO, DataQueryRequest, FieldType, ScopedVars } from '@grafana/data';
 import { AlarmsQueryHandlerCore } from 'datasources/alarms/query-type-handlers/AlarmsQueryHandlerCore';
-import { defaultAlarmsTrendQuery } from 'datasources/alarms/constants/DefaultQueries.constants';
+import { defaultAlarmTrendQuery } from 'datasources/alarms/constants/DefaultQueries.constants';
 import { Alarm, AlarmTransitionSeverityLevel, AlarmTransitionType, TransitionInclusionOption } from 'datasources/alarms/types/types';
-import { AlarmsTrendQuery, AlarmTransitionEvent, AlarmTrendSeverityLevelLabel, AlarmWithNumericTimeInTransitions } from 'datasources/alarms/types/AlarmsTrend.types';
+import { AlarmTrendQuery, AlarmTransitionEvent, AlarmTrendSeverityLevelLabel, AlarmWithNumericTimeInTransitions } from 'datasources/alarms/types/AlarmTrend.types';
 
-export class AlarmsTrendQueryHandler extends AlarmsQueryHandlerCore {
-  public readonly defaultQuery = defaultAlarmsTrendQuery;
+export class AlarmTrendQueryHandler extends AlarmsQueryHandlerCore {
+  public readonly defaultQuery = defaultAlarmTrendQuery;
   private readonly countKey = 'Count';
 
-  public async runQuery(query: AlarmsTrendQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
+  public async runQuery(query: AlarmTrendQuery, options: DataQueryRequest): Promise<DataFrameDTO> {
     const { start, end, intervalMs } = this.extractTimeParameters(options);
     const startTime = start.getTime();
     const endTime = end.getTime();